Unify files dealing with TLS signature schemes
The handling was split into multiple files compiled with different
build tags depending on the Go version, which allowed for use of
Ed25519 as the signature scheme for dcrwallet RPC certificates while
still supporting builds on Go 1.12.  This is no longer necessary now
that pre-1.13 Go support has been removed.
